Move home year view to route instead of just widget
