+++ lib/views/home/local_widgets/home_sheet.dart
@@ -10,8 +10,8 @@ import 'package:storypad/views/archives/archives_view.dart';
 import 'package:storypad/views/home/home_view_model.dart';
 import 'package:storypad/views/home/local_widgets/backup_tile.dart';
 import 'package:storypad/views/home/local_widgets/community_tile.dart';
-import 'package:storypad/views/home/local_widgets/home_years_view.dart';
 import 'package:storypad/views/home/local_widgets/language_tile.dart';
+import 'package:storypad/views/home/years_view/home_years_view.dart';
 import 'package:storypad/views/library/library_view.dart';
 import 'package:storypad/views/search/search_view.dart';
 import 'package:storypad/views/tags/tags_view.dart';
