+++ lib/views/home/years_view/new/new_year_view_model.dart
@@ -1,0 +1,10 @@
+import 'package:storypad/widgets/view/base_view_model.dart';
+import 'new_year_view.dart';
+
+class NewYearViewModel extends BaseViewModel {
+  final NewYearRoute params;
+
+  NewYearViewModel({
+    required this.params,
+  });
+}
