+++ lib/views/tags/edit/edit_tag_content.dart
@@ -9,12 +9,6 @@ class _EditTagContent extends StatelessWidget {
   Widget build(BuildContext context) {
     return SpTextInputsPage(
       appBar: AppBar(
-        actions: [
-          IconButton(
-            icon: Icon(Icons.show_chart),
-            onPressed: () => MessengerService.of(context).showSnackBar("message"),
-          ),
-        ],
         title: viewModel.tag != null ? Text(tr("page.edit_tag.title")) : Text(tr("page.new_tag.title")),
       ),
       fields: [
