+++ lib/views/tags/edit/edit_tag_view.dart
@@ -1,5 +1,4 @@
 import 'package:easy_localization/easy_localization.dart';
-import 'package:storypad/core/services/messenger_service.dart';
 import 'package:storypad/widgets/view/view_model_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:storypad/core/databases/models/tag_db_model.dart';
